Move mail logic to sso app
