+++ sso/mail.py
@@ -1,0 +1,10 @@
+from stregsystem.mail import send_template_mail
+
+
+def send_fcode_mail(member, fcode, redirect_url):
+    send_template_mail(
+        member,
+        "send_otp.html",
+        {**vars(member), 'fcode': fcode, 'redirect_url': redirect_url},
+        "Stregsystem FFO - F-Code",
+    )
