+++ sso/templates/mail/send_otp.html
@@ -5,7 +5,6 @@
 
         Din F-kode er {{ fcode }}. Koden udløber om 5 minutter.
 
-        Du vil modtage en opfølgende mail, når login'et er gennemført.
         Hvis du mener der er sket en fejl eller du ikke har anmodet en F-kode, kan du se bort fra denne mail, en Fember kan have indtastet et forkert brugernavn.
 
         Mvh,<br>
