+++ stregsystem/mail.py
@@ -33,15 +33,6 @@ def send_payment_mail(member, amount, mobilepay_comment):
     )
 
 
-def send_fcode_mail(member, fcode, redirect_url):
-    send_template_mail(
-        member,
-        "send_otp.html",
-        {**vars(member), 'fcode': fcode, 'redirect_url': redirect_url},
-        "Stregsystem FFO - F-Code",
-    )
-
-
 data_sent = {}
